Silence compiler warning
Refs trac:3796

This commit was SVN r29352.

The following Trac tickets were found above:
  Ticket 3796 --> https://svn.open-mpi.org/trac/ompi/ticket/3796

+++ ompi/communicator/comm_init.c
@@ -421,7 +421,7 @@ static void ompi_comm_destruct(ompi_communicator_t* comm)
     }
 
     /* mark this cid as available */
-    if ( MPI_UNDEFINED != comm->c_contextid &&
+    if ( MPI_UNDEFINED != (int)comm->c_contextid &&
          NULL != opal_pointer_array_get_item(&ompi_mpi_communicators,
                                              comm->c_contextid)) {
         opal_pointer_array_set_item ( &ompi_mpi_communicators,
